refactor(front): adjusts location of and deletes content components unrelated to markdown rendering
