+++ front/src/lib/components/posts/Footnotes.svelte
